+++ bindings/nanobind/CMakeLists.txt
@@ -41,6 +41,7 @@ set(PYWRAP_SOURCES
     algorithm/constrained-dynamics.cpp
     algorithm/constrained-dynamics-derivatives.cpp
     algorithm/impulse-dynamics.cpp
+    algorithm/impulse-dynamics-derivatives.cpp
     algorithm/contact-dynamics.cpp
     algorithm/contact-inverse-dynamics.cpp
     algorithm/contact-jacobian.cpp
